Comments above top level tag not supported
With the update to the comment handling code in support of python3,
comments above the top level tag would throw a parser error.  The
only fix for this is to move all comments inside the top level tag.

In the end, this is a better fix because on python2 this comments
were simply ignored and lost.  Now they will be preserved in both
python2 or python3

Signed-off-by: Jason Albert <albertj@us.ibm.com>